Update test data
The parts in multi compose are now named reasonably.

The package set repos are selected only once and not per variant. We
can't use repos with packages for specific arch only as that would
require more transformations during generating test data.

Lookasides for ResilientStorage part are added to point to Server part.

Signed-off-by: Lubomír Sedlář <lsedlar@redhat.com>